requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
cloudinary==1.40.0
decorator==5.1.1
dj-database-url==2.2.0
Django==5.0.7
django-appconf==1.0.6
django-avatar==8.0.0
django-registration-redux==2.13
django-widget-tweaks==1.5.0
dnspython==2.6.1
future==1.0.0
geocoder==1.38.1
gunicorn==22.0.0
idna==3.7
packaging==24.1
pillow==10.4.0
psycopg2-binary==2.9.9
python-decouple==3.8
ratelim==0.1.6
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.2
whitenoise==6.7.0
